try hacky fix

+++ test/test_ops_gradients.py
@@ -10,8 +10,9 @@ from torch.testing._internal.common_methods_invocations import op_db
 from torch.testing._internal.common_device_type import \
     (instantiate_device_type_tests, ops, OpDTypes)
 
-# TODO: fixme https://github.com/pytorch/pytorch/issues/68972
-torch.set_default_dtype(torch.float32)
+# To mitigate for ScalarxTensor type promotion
+# for complex scalar and float tensor.
+torch.set_default_dtype(torch.float64)
 
 # gradcheck requires double precision
 _gradcheck_ops = partial(ops, dtypes=OpDTypes.supported,
